Replaced vector with IIntervalOps protocol
The use of third and subsequent elements of vectors is a poor
substitute for protocols.

This commit ensures that data structures (types and maps) can behave
as intervals that can be operated on directly using tick's interval
operations, with the minimal constraints that such structures can be
narrowed and combined.