Alter semantics of time function to  mean local-time
There was some ambiguity as to what the `tick.core/time` function was
for (parsing, coercion, construction) and what the difference between
`tick.core/time` and `tick.core/now` was.

We now formally align 'time' with java.time.LocalTime. Think about it
like this: when you ask someone for the 'time', as in "what's the
time", the java.time.LocalTime (e.g. "4pm") is an acceptable answer.