Constraint semantics of interval constructor.
This is a much needed tidy up to avoid the overloading of the
construction of an interval, which leads to ambiguity, confusion and
potential bugs.

To add back some convenience, `temporal-value` is added to `tick.core`
which coerces java.lang.String and java.util.Date instances to
canonical tick values, allowing us to use these instances in tests.