Configure print-method for AtomClock
Fix #89

The intention of the old code was to make AtomClock print at the REPL as
if it were a map (I think?).  Unfortunately, it also messed with ALL
things which implement both IPersistentMap and IDeref, which then
impacts other types which do that too (e.g. discljord has a
DerefablePromiseChannel which does this).